+++ CHANGELOG.md
@@ -10,6 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
+
+## [Unreleased]
+
+### Added
+
+### Changed
 * `ExpressionBuilder`
     * `::neq` changed to `::ne`
     * `::lte` changed to `::le`
@@ -511,7 +524,8 @@ it's standard server error, not uri error.
 ### Security
 
 ---
-[Unreleased]: https://gitlab.com/bednic/json-api/compare/4.2.4...4.x
+[Unreleased]: https://gitlab.com/bednic/json-api/compare/5.0.0...5.x
+[4.2.4]: https://gitlab.com/bednic/json-api/compare/4.2.4...5.0.0
 [4.2.4]: https://gitlab.com/bednic/json-api/compare/4.2.3...4.2.4
 [4.2.3]: https://gitlab.com/bednic/json-api/compare/4.2.2...4.2.3
 [4.2.2]: https://gitlab.com/bednic/json-api/compare/4.2.1...4.2.2